docs: renumber Browser Use guide to 09 and switch to a single-file example
