--- docs/03_guides/code/browser_use_project/my_actor/__main__.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-import asyncio
-
-from .main import main
-
-if __name__ == '__main__':
-    asyncio.run(main())
